+++ blessed/keyboard.py
@@ -3,6 +3,8 @@
 # std imports
 import curses.has_key
 import curses
+import time
+import re
 
 # 3rd party
 import six
@@ -218,8 +220,8 @@ def resolve_sequence(text, mapper, codes):
     determine that ``xxx`` remains unresolved.
 
     :arg text: string of characters received from terminal input stream.
-    :arg OrderedDict mapper: an OrderedDict of unicode multibyte sequences,
-        such as u'\x1b[D' paired by their integer value (260)
+    :arg OrderedDict mapper: unicode multibyte sequences, such as ``u'\x1b[D'``
+        paired by their integer value (260)
     :arg dict codes: a :type:`dict` of integer values (such as 260) paired
         by their mnemonic name, such as ``'KEY_LEFT'``.
     :rtype: Keystroke
@@ -228,6 +230,87 @@ def resolve_sequence(text, mapper, codes):
         if text.startswith(sequence):
             return Keystroke(ucs=sequence, code=code, name=codes[code])
     return Keystroke(ucs=text and text[0] or u'')
+
+
+def _time_left(stime, timeout):
+    """
+    Return time remaining since ``stime`` before given ``timeout``.
+
+    This function assists determining the value of ``timeout`` for
+    class method :meth:`~.Terminal.kbhit` and similar functions.
+
+    :arg float stime: starting time for measurement
+    :arg float timeout: timeout period, may be set to None to
+       indicate no timeout (where None is always returned).
+    :rtype: float or int
+    :returns: time remaining as float. If no time is remaining,
+       then the integer ``0`` is returned.
+    """
+    if timeout is not None:
+        if timeout == 0:
+            return 0
+        return max(0, timeout - (time.time() - stime))
+
+
+def _read_until(term, pattern, timeout):
+    """
+    Convenience read-until-pattern function, supporting :meth:`~.get_location`.
+
+    :arg blessed.Terminal term: :class:`~.Terminal` instance.
+    :arg float timeout: timeout period, may be set to None to indicate no
+        timeout (where 0 is always returned).
+    :arg str pattern: target regular expression pattern to seek.
+    :rtype: tuple
+    :returns: tuple in form of ``(match, str)``, *match*
+        may be :class:`re.MatchObject` if pattern is discovered
+        in input stream before timeout has elapsed, otherwise
+        None. ``str`` is any remaining text received exclusive
+        of the matching pattern).
+
+    The reason a tuple containing non-matching data is returned, is that the
+    consumer should push such data back into the input buffer by
+    :meth:`~.Terminal.ungetch` if any was received.
+
+    For example, when a user is performing rapid input keystrokes while its
+    terminal emulator surreptitiously responds to this in-band sequence, we
+    must ensure any such keyboard data is well-received by the next call to
+    term.inkey() without delay.
+    """
+    stime = time.time()
+    match, buf = None, u''
+
+    # first, buffer all pending data. pexpect library provides a
+    # 'searchwindowsize' attribute that limits this memory region.  We're not
+    # concerned about OOM conditions: only (human) keyboard input and terminal
+    # response sequences are expected.
+
+    while True:
+        # block as long as necessary to ensure at least one character is
+        # received on input or remaining timeout has elapsed.
+        ucs = term.inkey(timeout=_time_left(stime, timeout))
+        if ucs:
+            buf += ucs
+            # while the keyboard buffer is "hot" (has input), we continue to
+            # aggregate all awaiting data.  We do this to ensure slow I/O
+            # calls do not unnecessarily give up within the first 'while' loop
+            # for short timeout periods.
+            while True:
+                ucs = term.inkey(timeout=0)
+                if not ucs:
+                    break
+                buf += ucs
+
+        match = re.search(pattern=pattern, string=buf)
+        if match is not None:
+            # match
+            break
+
+        if timeout is not None:
+            if not _time_left(stime, timeout):
+                # timeout
+                break
+
+    return match, buf
 
 
 def _inject_curses_keynames():
@@ -240,7 +323,7 @@ def _inject_curses_keynames():
     curses module, and is called from the global namespace at time of
     import.
 
-    Though we may determine keynames and codes for keyboard input that
+    Though we may determine *keynames* and codes for keyboard input that
     generate multibyte sequences, it is also especially useful to aliases
     a few basic ASCII characters such as ``KEY_TAB`` instead of ``u'\t'`` for
     uniformity.