+++ blessed/tests/test_keyboard.py
@@ -596,6 +596,62 @@ def test_esc_delay_cbreak_prefix_sequence():
     assert 34 <= int(duration_ms) <= 45, duration_ms
 
 
+def test_get_location_0s():
+    "0-second get_location call without response."
+    @as_subprocess
+    def child():
+        term = TestTerminal(stream=six.StringIO())
+        stime = time.time()
+        y, x = term.get_location(timeout=0)
+        assert (math.floor(time.time() - stime) == 0.0)
+        assert (y, x) == (-1, -1)
+    child()
+
+
+def test_get_location_0s_under_raw():
+    "0-second get_location call without response under raw mode."
+    @as_subprocess
+    def child():
+        term = TestTerminal(stream=six.StringIO())
+        with term.raw():
+            stime = time.time()
+            y, x = term.get_location(timeout=0)
+            assert (math.floor(time.time() - stime) == 0.0)
+            assert (y, x) == (-1, -1)
+    child()
+
+
+def test_get_location_0s_reply_via_ungetch():
+    "0-second get_location call with response."
+    @as_subprocess
+    def child():
+        term = TestTerminal(stream=six.StringIO())
+        stime = time.time()
+        # monkey patch in an invalid response !
+        term.ungetch(u'\x1b[10;10R')
+
+        y, x = term.get_location(timeout=0.01)
+        assert (math.floor(time.time() - stime) == 0.0)
+        assert (y, x) == (10, 10)
+    child()
+
+
+def test_get_location_0s_reply_via_ungetch_under_raw():
+    "0-second get_location call with response under raw mode."
+    @as_subprocess
+    def child():
+        term = TestTerminal(stream=six.StringIO())
+        with term.raw():
+            stime = time.time()
+            # monkey patch in an invalid response !
+            term.ungetch(u'\x1b[10;10R')
+
+            y, x = term.get_location(timeout=0.01)
+            assert (math.floor(time.time() - stime) == 0.0)
+            assert (y, x) == (10, 10)
+    child()
+
+
 def test_keystroke_default_args():
     "Test keyboard.Keystroke constructor with default arguments."
     from blessed.keyboard import Keystroke